feat(alert)!: change arguments for alert
The old alert had received the initial message even when recover alert,
and the argument was the same order as log file but no latency record.
That was a bit difficult to understand.
By this commit, the alert receive always the latest status and message,
and the argument is completely the same as log file, includes latest latency data.
It's easier to understand than used to.